+++ torch_geometric/explain/metric/__init__.py
@@ -1,8 +1,9 @@
 from .basic import groundtruth_metrics
-from .fidelity import characterization_score, fidelity_curve_auc
+from .fidelity import fidelity, characterization_score, fidelity_curve_auc
 
 __all__ = classes = [
     'groundtruth_metrics',
+    'fidelity',
     'characterization_score',
     'fidelity_curve_auc',
 ]